# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Online Movie Hub API
  description: A simple API to do CRUD on users, book tickets, watch movies in online app
servers:
  - url: http://localhost:5000/
...
paths:
...
  /movie:
    get:
      summary: Lists all the Movies
      tags: [Movies]
      responses:
        "200":
          description: The list of movies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movies"
    post:
      summary: Creates a Movie Object.
      tags: [Movies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movies"
      responses:
        "201":
          description: Created Successfully.
  /movie/{movieID}:
    get:
      summary: Returns a Movie Object by ID.
      tags: [Movies]
      parameters:
        - name: movieID
          in: path
          required: true
          description: Autogenerated Movie ID by MongoDB
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movies"
    patch:
      tags: [Movies]
      summary: Updated Movie
      description: updates the movie object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Movies"
      produces:
        - "application/json"
      parameters:
        - name: movieID
          in: path
          description: object that needs to be updated
          required: true
          type: string
      responses:
        "400":
          description: Invalid movie ID supplied
        "404":
          description: movie not found
    delete:
      tags: [Movies]
      summary: Delete movie
      description: deletes the movie object
      produces:
        - "application/json"
      parameters:
        - name: movieID
          in: path
          description: The object ID that needs to be deleted
          required: true
          type: string
      responses:
        "400":
          description: Invalid movie ID supplied
        "404":
          description: movie not found
...
components:
  schemas:
...
    Movies:
      type: object
      required:
        - title
        - genre
        - imdb
        - image
        - duration
        - trailer
      properties:
        id:
          type: integer
          description: The auto-generated id of the movie.
        title:
          type: string
          description: The title of the movie.
        genre:
          type: string
          description: The category of movie.
        imdb:
          type: number
          description: The rating of the movie.
        image:
          type: string
          description: The poster of the movie.
        duration:
          type: string
          description: The running time of the movie.
        trailer:
          type: string
          description: The trailer of the movie.
      example:
        title: Spider Man No Way Home
        genre: Action
        imdb: 8.9
        image: https://images.thedirect.com/media/photos/FFBVo2XaMAA8ydw.jpg
        duration: 2 hr 28 min
        trailer: https://youtu.be/ZYzbalQ6Lg8
...